Don't turn the device off when receiving read errors
We assumed if there is read error, the device has been disconnected.
This may not be true. Let hotplugging code turn the device off if
the device is disconnected. Otherwise, just log an error.

However, we get 10 read errors when a device is disconnected before
it is actually removed from the driver. One log message shows all
we need to know.

Signed-off-by: Ping Cheng <pingc@wacom.com>