Initial changeset to produce Max's code from late April 2011
All unnecessary files have been removed (there were a lot of backup,
temporary, and already-built just lying around) and permissions have
been set to sane values. Although the diff for this commit is huge,
most of the it is due to the removal of almost all Linux kernel code.

Max's code was provided with a copy of the two third-party programs
"expresskeys-0.4.2" and "wacom-config-1.0b3.py". These have been
removed since they were never referenced by any of the build scripts,
and can be found online.