Fix in-proximity bit setting AND fix non-21UX2 packet handling
This patch should be split.

This (along with the prior patch) fixes a problem where only the
left buttons of the 21UX2 were usable. With the in-proximity bit
not being properly set, the driver was loosing them.

This also fixes the handling of non-21UX2 packets. The wrong bits
are bieng inspected in some cases, causing problems for the
e.g. 21UX.