Add option to disable pressure recalibration
If the initial pressure of a device is != 0 the driver recalibrates
the pressure range. This is to account for worn out devices.
The downside is that when the user hits the tablet very hard the
initial pressure reading may be unequal to zero even for a perfectly
good pen. If the consecutive pressure readings are not higher than
the initial pressure by a threshold no button event will be generated.
This option allows to disable the recalibration.

Signed-off-by: Egbert Eich <eich@suse.de>
Signed-off-by: Ping Cheng <pingc@wacom.com>