WIP: Properly handle ncurses include at "ncurses/ncurses.h"
The configure script currently detects if the ncurses header is located
at "ncurses/ncurses.h", but compilation will still fail since we use
a plain "ncurses.h" include and make no attempt to update the include
path. This commit detects when the include is actually located under
a subdirectory and ensures we include that file instead.